Вынес получение изображения отдельно в библиотеку
