+++ Homework/TagsCloud.Console/IConsoleUI.cs
@@ -1,7 +1,9 @@
-﻿namespace TagsCloud.Console
+﻿using TagsCloudContainer;
+
+namespace TagsCloud.Console
 {
     public interface IConsoleUI
     {
-        public void Run(IAppSettings settings);
+        public void Run(IAppSettings appSettings,ITagCloudSettings tagCloudsettings);
     }
 }
