+++ Homework/TagsCloudContainer/ITagCloud.cs
@@ -1,0 +1,10 @@
+﻿using System.Collections.Generic;
+using System.Drawing;
+
+namespace TagsCloudContainer
+{
+    public interface ITagCloud
+    {
+        public Bitmap LayDown(IEnumerable<string> words, ITagCloudSettings settings);
+    }
+}
